+++ server/src/main/kotlin/com/github/csc3380fall2024/team16/Application.kt
@@ -1,5 +1,11 @@
 package com.github.csc3380fall2024.team16
 
+import com.github.csc3380fall2024.team16.plugins.configureDatabase
+import com.github.csc3380fall2024.team16.plugins.configureExceptions
+import com.github.csc3380fall2024.team16.plugins.configureRouting
+import com.github.csc3380fall2024.team16.plugins.configureSecurity
+import com.github.csc3380fall2024.team16.plugins.configureSerialization
+import com.github.csc3380fall2024.team16.plugins.configureValidation
 import io.ktor.server.application.Application
 import io.ktor.server.engine.embeddedServer
 import io.ktor.server.netty.Netty
