+++ server/src/main/kotlin/com/github/csc3380fall2024/team16/plugins/Exceptions.kt
@@ -1,4 +1,4 @@
-package com.github.csc3380fall2024.team16
+package com.github.csc3380fall2024.team16.plugins
 
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.Application
