+++ server/src/main/kotlin/com/github/csc3380fall2024/team16/plugins/Validation.kt
@@ -1,5 +1,7 @@
-package com.github.csc3380fall2024.team16
+package com.github.csc3380fall2024.team16.plugins
 
+import com.github.csc3380fall2024.team16.LoginDto
+import com.github.csc3380fall2024.team16.RegisterDto
 import io.ktor.server.application.Application
 import io.ktor.server.application.install
 import io.ktor.server.plugins.requestvalidation.RequestValidation
